add support for setting environment variables for simulation
Currently we translate the model using OMPython and then run the executable
in a different process which has a different environment. This is a problem
if the user sets some environment variable in omc that is suppose to be used
by the simulation executable.

This commit adds support for specifying some environment [variable, value]
pairs in environmentSimulation in JSON. We also support $libraryLocation
in the value to be replaced by the actual location of the library on the
disk.